test-mcp/requirements.txt
requests>=2.28.0
atlassian-python-api>=3.41.0
asyncio